AI Vacuum C#
This program uses Boolean and String data type variables to determine if room 'A' or 'B' is clean or not. If the current room is dirty, the AI vacuum will clean it. Then the state of that current room will be changed to "clean" and the vacuum will move onto the next room. This program also randomizes which room is dirty and randomizes which room to check first.