Patrolling AI Enemy C#
This script is meant to be placed on a game object inside of the game engine Unity. This script can be attached to a game object to give it a patrolling enemy behavior. The enemy patrols around a certain area based on an array of way points placed in the game environment.